refactor(frontend): Fields with zod validation
